Restructure thrFiringRemainderAlg Types.h to pure C
